requirements.txt
python-telegram-bot==20.7
pandas==2.1.4
pandas-ta==0.3.14b
requests==2.31.0
python-dotenv==1.0.0
apscheduler==3.10.4
aiohttp==3.9.1
numpy==1.26.3
